get that X on settings modal to behave
