+++ corpan/corpan-app/src/components/MainExperience.tsx
@@ -17,35 +17,12 @@ import { createVoiceTTS } from "@/util/speak";
 import { useTranslation } from "react-i18next";
 
 import { isRTL, toCamelCase } from "@/util/convert";
-import { isAndroid } from "@/util/browser";
-
-
-export function getPlatformBottomPadding() {
-    if (/iPhone|iPad|iPod|iOS/i.test(navigator.userAgent)) {
-        return 180;
-    } if (/Android/i.test(navigator.userAgent)) {
-        return 195;
-    }
-    return 180;
-}
-
-export function getPlatformTopPaddingButtons() {
-    if (/iPhone|iPad|iPod|iOS/i.test(navigator.userAgent)) {
-        return 55;
-    } if (/Android/i.test(navigator.userAgent)) {
-        return 27;
-    }
-    return 0;
-}
-
-export function getPlatformTopPaddingTranslations() {
-    if (/iPhone|iPad|iPod|iOS/i.test(navigator.userAgent)) {
-        return 150;
-    } if (/Android/i.test(navigator.userAgent)) {
-        return 125;
-    }
-    return 75;
-}
+import {
+    getPlatformBottomPadding,
+    getPlatformTopPaddingButtons,
+    getPlatformTopPaddingTranslations,
+    isAndroid,
+} from "@/util/browser";
 
 // // Even lamer but still fine
 // const paddingAdjustMap: Record<string, number> = {
